feat(tabs): remove internal support for history and url manipulation
Removes baked in support for manipulation of window location as implementation is extremely limited
and implementing the a feature to override it will be difficult to maintain. Consumers wishing to
have this functionality should utilise the `onTabChange` and `selectedTabId` props and implement
their own `history` implementation.

BREAKING CHANGE: Internal support for updating window location has been removed from the `Tabs`
component, the `setLocation` prop has also been removed as it serves no purpose beyond serving as a
guard for this functionality. Consumers who wish to have this functionality should utilise the
`onTabChange` and `selectedTabId` props with whatever `history` implementation they want to.

closes: #3520